feat(Topup): Add event driven service for store transaction

+++ microservices/input-harga-service/internal/dto/dto.go
@@ -1,10 +1,10 @@
 package dto
 
 type InputHargaRequest struct {
-	AdminId      string  `json:"admin_id"`
-	ReffID       string  `json:"reff_id"`
-	HargaTopup   float64 `json:"harga_topup"`
-	HargaBuyback float64 `json:"harga_buyback"`
+	AdminId      string `json:"admin_id"`
+	ReffID       string `json:"reff_id"`
+	HargaTopup   int64  `json:"harga_topup"`
+	HargaBuyback int64  `json:"harga_buyback"`
 }
 
 type InputHargaResponse struct {
