+++ microservices/topup-service/internal/dto/dto.go
@@ -3,7 +3,7 @@ package dto
 type TopupRequest struct {
 	ReffId     string
 	GoldWeight float64 `json:"gram"`
-	Price      float64 `json:"harga"`
+	Price      int64   `json:"harga"`
 	Norek      string  `json:"norek"`
 }
